Se agrega CRUD de docnete y validaciones

+++ src/main/java/com/grup14/luterano/repository/DocenteRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface DocenteRepository extends JpaRepository<Docente,Integer> {
+public interface DocenteRepository extends JpaRepository<Docente,Long> {
     Optional<Docente> findByEmail(String email);
     Optional<Docente> findByDni(String dni);
 }
